# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: File Workshop API
  version: 0.1.0
  description: File Workshop V1.0 REST API 工作契约。业务 API 使用 /api/v1，平台健康检查为部署探针例外。
servers:
  - url: /
...
paths:
...
  /api/v1/audit/events:
    get:
      operationId: listAuditEvents
      tags: [Audit]
      summary: List audit events.
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - name: dateFrom
          in: query
          required: true
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          required: true
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - name: eventType
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 128 }
        - name: riskLevel
          in: query
          required: false
          schema: { $ref: "#/components/schemas/AuditRiskLevel" }
        - name: actorType
          in: query
          required: false
          schema: { $ref: "#/components/schemas/AuditActorType" }
        - name: actorId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: resourceType
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 32 }
        - name: resourceId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: result
          in: query
          required: false
          schema: { $ref: "#/components/schemas/AuditResult" }
        - name: requestId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Audit event list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventListResponse"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/audit/events/{auditEventId}:
    get:
      operationId: getAuditEvent
      tags: [Audit]
      summary: Get audit event detail.
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/AuditEventIdPath"
        - name: partitionDate
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Audit event detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventResponse"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/audit/integrity:
    get:
      operationId: getAuditIntegrity
      tags: [Audit]
      summary: List audit hash chain heads.
      security:
        - bearerAuth: []
        - sessionCookie: []
      parameters:
        - name: dateFrom
          in: query
          required: true
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          required: true
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/AuditChainStatus" }
      responses:
        "200":
          description: Audit integrity chain heads.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditIntegrityResponse"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/audit/integrity/verify:
    post:
      operationId: verifyAuditIntegrity
      tags: [Audit]
      summary: Verify one audit hash chain.
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyAuditIntegrityRequest"
      responses:
        "200":
          description: Audit chain verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditIntegrityVerificationResponse"
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 短期访问 JWT。浏览器不应将其持久化到 LocalStorage。
    sessionCookie:
      type: apiKey
      in: cookie
      name: file_workshop_access
      description: 浏览器短期访问 JWT Cookie，使用 HttpOnly 和 SameSite；生产环境必须使用 Secure。
    refreshCookie:
      type: apiKey
      in: cookie
      name: file_workshop_refresh
      description: 只用于刷新和注销的轮换 Refresh Token Cookie，使用 HttpOnly 和 SameSite；生产环境必须使用 Secure。
  parameters:
    PageQuery:
      name: page
      in: query
      required: false
      description: 从 1 开始的页码。
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      required: false
      description: 每页数量，最大 200。
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    AuditEventIdPath:
      name: auditEventId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  headers:
    RequestId:
      description: 本次请求的稳定追踪标识。
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    InvalidRequest:
      description: 请求 JSON 或字段不符合契约。
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthRequired:
      description: 未提供有效认证信息、访问令牌过期或会话已经失效。
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: 当前用户没有执行该管理操作的权限。
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 指定资源不存在或对当前用户不可见。
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: 唯一值、幂等键、用户状态或 rowVersion 发生冲突。
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    AuditRiskLevel:
      type: string
      enum: [NORMAL, HIGH, CRITICAL]
    AuditActorType:
      type: string
      enum: [USER, SYSTEM, AGENT, MIGRATION, SERVICE]
    AuditResult:
      type: string
      enum: [SUCCESS, FAILURE, DENIED]
    AuditSourceChannel:
      type: string
      enum: [WEB, API, WEBDAV, AGENT, MIGRATION, SYSTEM]
    AuditChainStatus:
      type: string
      enum: [ACTIVE, SEALED, INVALID]
    AuditEvent:
      type: object
      additionalProperties: false
      required: [auditEventId, eventType, riskLevel, actorType, action, result, sourceChannel, requestId, metadataSchemaVersion, metadata, partitionDate, createdAt]
      properties:
        auditEventId: { type: string, format: uuid }
        eventType: { type: string, minLength: 1, maxLength: 128 }
        riskLevel: { $ref: "#/components/schemas/AuditRiskLevel" }
        actorType: { $ref: "#/components/schemas/AuditActorType" }
        actorId: { type: string, format: uuid }
        actorDisplayName: { type: string, maxLength: 256 }
        actorEmployeeNo: { type: string, maxLength: 128 }
        effectiveRole: { type: string, maxLength: 32 }
        adminDelegationId: { type: string, format: uuid }
        shareId: { type: string, format: uuid }
        resourceType: { type: string, maxLength: 32 }
        resourceId: { type: string, format: uuid }
        resourceName: { type: string, maxLength: 512 }
        spaceId: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        documentVersionId: { type: string, format: uuid }
        action: { type: string, minLength: 1, maxLength: 128 }
        result: { $ref: "#/components/schemas/AuditResult" }
        failureCode: { type: string, maxLength: 64 }
        sourceChannel: { $ref: "#/components/schemas/AuditSourceChannel" }
        ipAddress: { type: string }
        userAgent: { type: string }
        requestId: { type: string, format: uuid }
        traceId: { type: string, maxLength: 128 }
        correlationId: { type: string, format: uuid }
        reason: { type: string }
        metadataSchemaVersion: { type: integer, minimum: 1 }
        metadata:
          type: object
          additionalProperties: true
        hashSchemaVersion: { type: integer, minimum: 1 }
        chainId: { type: string, maxLength: 128 }
        sequenceNumber: { type: integer, format: int64, minimum: 1 }
        previousHash: { type: string, format: byte }
        eventHash: { type: string, format: byte }
        partitionDate: { type: string, format: date }
        createdAt: { type: string, format: date-time }
    AuditEventListResponse:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total, requestId]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AuditEvent" }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, format: int64, minimum: 0 }
        requestId: { type: string }
    AuditEventResponse:
      type: object
      additionalProperties: false
      required: [event, requestId]
      properties:
        event: { $ref: "#/components/schemas/AuditEvent" }
        requestId: { type: string }
    AuditChainHead:
      type: object
      additionalProperties: false
      required: [chainId, partitionDate, lastSequenceNumber, lastEventId, lastHash, status, createdAt, updatedAt, rowVersion]
      properties:
        chainId: { type: string, minLength: 1, maxLength: 128 }
        partitionDate: { type: string, format: date }
        lastSequenceNumber: { type: integer, format: int64, minimum: 0 }
        lastEventId: { type: string, format: uuid }
        lastHash: { type: string, format: byte }
        batchRoot: { type: string, format: byte }
        anchorLocation: { type: string, maxLength: 1024 }
        status: { $ref: "#/components/schemas/AuditChainStatus" }
        verifiedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        rowVersion: { type: integer, format: int64, minimum: 1 }
    AuditIntegrityResponse:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total, requestId]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AuditChainHead" }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 200 }
        total: { type: integer, format: int64, minimum: 0 }
        requestId: { type: string }
    VerifyAuditIntegrityRequest:
      type: object
      additionalProperties: false
      required: [chainId, partitionDate]
      properties:
        chainId: { type: string, minLength: 1, maxLength: 128 }
        partitionDate: { type: string, format: date }
    AuditIntegrityVerificationResponse:
      type: object
      additionalProperties: false
      required: [chainId, partitionDate, verified, checkedEvents, requestId]
      properties:
        chainId: { type: string }
        partitionDate: { type: string, format: date }
        verified: { type: boolean }
        checkedEvents: { type: integer, minimum: 0 }
        failureReason: { type: string }
        requestId: { type: string }
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
        message:
          type: string
          minLength: 1
          maxLength: 512
        requestId:
          type: string
          minLength: 1
          maxLength: 128
        details:
          type: object
          additionalProperties: true